2opt uses knn

+++ src/TSP.h
@@ -22,7 +22,7 @@ public:
 
     static vector<int> travel_cw_seq(const vector<pair<double, double>> &cities, Grid<int> &distances);
 
-    static vector<int> local_2opt(vector<int> tour, Grid<int> &distances, Grid<int> &knearest,
+    static vector<int> local_2opt(vector<int> tour, Grid<int> &distances, Grid<uint16_t> &knearest,
                                   time_point<system_clock, duration<long, ratio<1, 1000000000>>> *deadline);
 
     static int tour_distance(const vector<pair<double, double>> &cities, vector<int> tour);
